+++ examples/example-expo/e2e/mock-server.ts
@@ -1,6 +1,5 @@
 import express, { Request, Response } from 'express'
 import bodyParser from 'body-parser'
-import { LZString } from 'posthog-core'
 const PORT = process.env.PORT || 8000
 
 export interface MockRequest {
@@ -18,18 +17,11 @@ export const createMockServer = (): [any, jest.Mock<MockRequest, any>] => {
   let httpMock = jest.fn()
 
   const handleRequest = (req: Request, res: Response) => {
-    let body
-    try {
-      body = req.body.compression === 'lz64' ? JSON.parse(LZString.decompressFromBase64(req.body.data) || '') : req.body
-    } catch (e) {
-      console.error(e)
-      body = 'error'
-    }
     const data: MockRequest = {
       method: req.method,
       path: req.path,
       headers: req.headers,
-      body: body,
+      body: req.body,
       params: req.params,
     }
     res.json(httpMock(data) || { status: 'ok' })