+++ posthog-web/CHANGELOG.md
@@ -1,5 +1,12 @@
 # Next
 
+# 4.0.0 - 2025-06-10
+
+## Removed
+
+1. Remove `captureMode` in favor of `json` capture mode only
+2. Remove deprecated `personProperties` and `groupProperties` in favor of `setPersonPropertiesForFlags` and `setGroupPropertiesForFlags`
+
 # 3.6.0 – 2025-06-05
 
 ## Added